Basic support for reverse proxys/CDNs/Docker/etc
In some situations REMOTE_ADDR might not point to the visitors address but the upstream proxy server. This leads to WP-Statistics only recording the same (server) IP over and over again. This change checks for the presence of typical headers proxys use to forward the originators IP and - if present - use these instead.

Note: Since these headers are not standardized this might still not work - making this section configurable would probably be a better idea in the long run