Fix Real-time stats column header width
This forces a width for the header of the Real-time stats column. If there's a lot of columns selected to be shown, the header text would squish down to be vertical text. I first discovered this issue with a WooCommerce store while browsing the Products page in WP backend which had a lot of columns selected to be shown. With this CSS rule addition I was able to fix the header for my WooCommerce store.