Update class-wp-statistics-shortcode.php
Now when you use [wpstatistics stat=abbreviated], it will show only the symbol (e.g. 'K', 'M', 'B') and hide the numbers. If you use any other value for stat, it will display numbers formatted in shorthand notation.

Make sure you have added the code for the formatNumber function and the modification in the shortcodes method as shown above. If you continue to experience issues or receive specific errors, please share the error message so I can provide you with a more specific solution.

[wpstatistics stat=visits time=total format=abbreviated]

In this example, the wpstatistics shortcode is used to get the number of visits and the format attribute is set to abbreviated to format the number in abbreviated notation.